Fix: task3

+++ src/Main.java
@@ -32,7 +32,7 @@ public class Main {
 
     public static double averageOfArray(int[] array, int daysPerMounth) {
         double result = 0.0;
-        return result = sumOfArray(array) / daysPerMounth;
+        return result = (double)sumOfArray(array) / daysPerMounth;
     }
 
     static void task1() {
